+++ ts/core/plugin.ts
@@ -1,7 +1,11 @@
-import { PLUGIN_LOGGER } from './private-loggers';
 import { CaretSelection } from './tool/dom-tools';
+import { PLUGIN_LOGGER } from './private-loggers';
 import { pp } from './tool/string';
 
+/**
+ * Contains mutable state for the editor.
+ * It is passed around plugins and mutated to edit the editor state.
+ */
 export class EditorState {
   constructor(
     private _content: string,
@@ -16,6 +20,15 @@ export class EditorState {
     return this._selection;
   }
 
+  /**
+   * Inserts text at a specified position.
+   * Ex.:
+   * ```typescript
+   * state.content; // "Hello World!"
+   * state.insertText("Wonderful ", 6);
+   * state.content; // "Hello Wonderful World!"
+   * ```
+   */
   public insertText(text: string, position: number) {
     if (position < 0 || position > this._content.length) {
       PLUGIN_LOGGER.ERROR(pp`Cannot insert text at position ${position}`);
@@ -25,6 +38,15 @@ export class EditorState {
     this._content = old.slice(0, position) + text + old.slice(position);
   }
 
+  /**
+   * Deletes a length of text at a specified position.
+   * Ex.:
+   * ```typescript
+   * state.content; // "Hello World!"
+   * state.deleteText(6, 6);
+   * state.content; // "Hello!"
+   * ```
+   */
   public deleteText(position: number, length: number) {
     if (position < 0 || position > this._content.length) {
       PLUGIN_LOGGER.ERROR(pp`Cannot delete text at position ${position}`);
@@ -45,7 +67,30 @@ export class EditorState {
   }
 }
 
+/**
+ * The provider for code injection and plugin creation, used for creating new plugins.
+ * Let's create one now!
+ * Start by sub-classing PluginProvider:
+ * ```typescript
+ *  class MyPlugin extends PluginProvider {
+ *
+ * }
+ * ```
+ * Next override any functions you need. Let's override onKeyPressed, and log a
+ * message when the user types an "a":
+ * ```typescript
+ * //...
+ * override onKeyPressed(key: string, state: EditorState) {  // We could totally omit state if we wanted.
+ *   if (key == "a") {
+ *     CLIENT_LOGGER.INFO(`"A" key pressed!`);
+ *   }
+ * }
+ * //...
+ * ```
+ * Now you just create a new {@link Plugin} and register it, and you see the messages in the console!
+ */
 export abstract class PluginProvider {
+  public onInitialize() {}
   public onKeyPressed(key: string, state: EditorState) {}
 }
 
@@ -53,6 +98,11 @@ export class PluginConfig {
   constructor({}) {}
 }
 
+/**
+ * A plugin sourc can use to inject code into the editor.
+ * To make one all you need is an instance of a subclass of {@link PluginProvider}
+ * and a {@link PluginConfig} object.
+ */
 export class SourcPlugin {
   constructor(
     public readonly provider: PluginProvider,