+++ ts/core/private-loggers.ts
@@ -1,13 +1,7 @@
-import { Logger, ConsoleLogStrategy } from './logger';
+import { Logger } from './logger';
 
-export const DOM_TOOLS_LOGGER = new Logger()
-  .withName('dom-tools.ts')
-  .withStrategy(ConsoleLogStrategy);
+export const DOM_TOOLS_LOGGER = new Logger().withName("dom-tools.ts");
 
-export const PLUGIN_LOGGER = new Logger()
-  .withName('plugin.ts')
-  .withStrategy(ConsoleLogStrategy);
+export const PLUGIN_LOGGER = new Logger().withName("plugin.ts");
 
-export const CORE_LOGGER = new Logger()
-  .withName('SOURC_CORE')
-  .withStrategy(ConsoleLogStrategy);
+export const CORE_LOGGER = new Logger().withName("SOURC_CORE");