+++ CHANGELOG.md
@@ -7,12 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Documentation for every publicly accessible thing
+- findLineOf string utility function
+
+### Removed
+
+- Remove TextEditor.overlay since it's unimplemented and
+  its functionality will be replicated in a future plugin anyway.
+- Remove TextEditor#setName since TextEditor.LOGGER#withName is good enough.
+
 ## [0.0.1] - 8/27/2021
 
 ### Added
 
 - Plugin registry
 - State handlers
+- Plugin#onKeyPressed
 
 ### Core Plugin